HAWKULAR-920 Refactor chart types out to individual files. Also split features into their own where they can be used by several chart types. First set of charts to be refactored is line, area, scatter and scatter-line. More to come.
