+++ index.html
@@ -70,8 +70,8 @@
 
   <!-- build:js scripts/vendor.js -->
   <!-- bower:js -->
-  <script src="libs/jquery/dist/jquery.js"></script>
   <script src="libs/angular/angular.js"></script>
+  <script src="libs/jquery/dist/jquery.js"></script>
   <script src="libs/d3/d3.js"></script>
   <script src="libs/moment/moment.js"></script>
   <script src="libs/d3-tip/index.js"></script>
